build: use setuptools-git-versioning for versioning (#25)
* build: use setuptools-git-versioning for versioning

The versioneer file here was never updated for 3.12.  Instead of
doing that now, I think setuptools-git-versioning will be a
better way of doing it, since there's no in-repo code to fall
out of date anymore.

I've used setuptools-git-versioning already in other things (see
`alpenhorn-chime`) and it seems like less of a burden than versioneer
has been.