refactor(util): Delete alpenhorn-1 import routines (#28)
This deletes all the routines used by alpenhorn-1 for file
detection and registration from the `util` module.  This
functionality was all re-implemented for alpenhorn-2 in #21
and https://github.com/chime-experiment/alpenhorn-chime/pull/1

All the file regexes now live in the `FileType` table in the database.

This could all just stay here and languish as unused dead code
without interfering, I suppose, but I think it's better to delete
it all to prevent future code devs by being misled by its presence.

I really don't understand why this code ended up here instead of
just being part of alpenhorn-1.  Perhaps it was simply missed when
alpenhorn was first split off from `ch_util` and then moved here when
this repository was similarly blebbed off.  Or perhaps there were plans
for some sort of non-alpenhorn-y way of adding files to the data index
that never came to fruition.  Whatever the reasoning, in reality, it's
never been referenced by anything other than alpenhorn-1.

**Note**: We are still running an alpenhorn-1 instance on niagara to
manage our HPSS archive.  However we shouldn't be importing new files
there, so while that instance _could_ call into this code, in practice
it never will.

I'd say if we need to redeploy the instance on niagara (and we don't
want to upgrade to alpenhorn-2 at that time), then a better solution
would be to disable the import code in that deployment's instance
instead of keeping this code around here so it could be referenced but
never used.